src/processor/requirements.txt
pyspark==3.3.1
pymongo==4.3.3
python-dotenv==1.0.0 